requirements.txt
langgraph[all]
langchain-core
langchain-openai
langchain-community
python-dotenv
flask
openai
ipython
graphviz 